feat(edusharing): create a state worker to correctly handle when deployment is done by edusharing
And minor fix
